all comments in same style + renamed random alogirithm class to Randomize

+++ code/classes/connection.py
@@ -1,12 +1,12 @@
 class Connection:
     """ Represents a connection between two train stations. """
 
-    def __init__(self, trajectory_number: int, station_1: str, station_2: str, duration: int) -> None:
+    def __init__(self, trajectory_number: int, station_1: str, station_2: str, duration: float) -> None:
         """ Creates a Connection object. """
         self.connection_number: int = trajectory_number
         self.station_1: str = station_1
         self.station_2: str = station_2
-        self.duration: int = duration
+        self.duration: float = duration
 
     def get_destination_station(self, departure_station: str) -> str:
         """ Given one station, returns the other end of the connection. """
